% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geom4PL.R
\name{stat_4PL}
\alias{stat_4PL}
\alias{geom_4PL}
\title{4-parameter logistic regression (4PL)}
\usage{
stat_4PL(
  geom = GeomLine,
  position = position_identity(),
  fct = drc::LL.4(names = c("Slope", "Lower Limit", "Upper Limit", "EC50")),
  ...
)

geom_4PL(
  mapping = NULL,
  data = NULL,
  stat = Stat4PL,
  position = position_identity(),
  fct = drc::LL.4(names = c("Slope", "Lower Limit", "Upper Limit", "EC50"))
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[drc:drm]{drc::drm}}
  \describe{
    \item{\code{fct}}{a list with three or more elements specifying the non-linear
  function, the accompanying self starter function, the names of the parameter in the non-linear function and, 
  optionally, the first and second derivatives as well as information used for calculation of ED values.
  Currently available functions include, among others, the four- and five-parameter log-logistic models 
  \code{\link[drc]{LL.4}}, \code{\link[drc]{LL.5}} and the Weibull model \code{\link[drc]{W1.4}}. Use
  \code{\link[drc]{getMeanFunctions}} for a full list.}
  }}
}
\description{
This stat will fit a 4PL line using \code{drc::drm()}.
The x data to be used will not be transformed.
In other word, the fitted model will not be influenced by \code{scale_x_()} functions.
The stat layer implement the regression model,
while the geom layer inherits from geom_line to make the plot.
}
