% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdose-package.R
\docType{package}
\name{ggdose-package}
\alias{ggdose}
\alias{ggdose-package}
\title{ggdose: Plot dose response curve}
\description{
Plot dose-response curve using the underlying data from ggplot.
}
\author{
\strong{Maintainer}: Yuhao Min \email{min.yuhao@mayo.edu}

}
\keyword{internal}
